refactor: move power_unit data
